+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.14.0]
+
+### Added
+
+- Added support for creating egress private endpoints via the
+  `cockroach_egress_private_endpoint` resource.
+
 ## [1.13.1]
 
 ### Fixed
@@ -203,7 +210,7 @@ removed from the terraform config
 
 ## Added
 
-- The [cockroach_api_key](https://registry.terraform.io/providers/cockroachdb/cockroach/latest/docs/resources/api_key) 
+- The [cockroach_api_key](https://registry.terraform.io/providers/cockroachdb/cockroach/latest/docs/resources/api_key)
   resource was added.
 
 - The [cockroach_service_account](https://registry.terraform.io/providers/cockroachdb/cockroach/latest/docs/resources/service_account)
@@ -297,7 +304,7 @@ removed from the terraform config
 
 ### Added
 
-- New folder resource and the new `parent_id` for clusters support users managing nested resources. 
+- New folder resource and the new `parent_id` for clusters support users managing nested resources.
 Folders is in [limited access](https://www.cockroachlabs.com/docs/v23.1/cockroachdb-feature-availability).
 
 ## [1.0.0] - 2023-08-10